#188 [MOVE] LoadingSpinner를 pages -> components 폴더로 이동
