+++ src/pages/onboarding/InviteLoading.tsx
@@ -4,7 +4,7 @@ import { useEffect } from 'react';
 import { useNavigate, useParams, useSearchParams } from 'react-router-dom';
 import { useLocalStorage } from '../../hooks/useLocalStorage';
 import { LOCAL_STORAGE_KEY } from '../../constants/key';
-import LoadingSpinner from '../LoadingSpinner';
+import LoadingSpinner from '../../components/LoadingSpinner';
 
 const InviteLoading = () => {
   const { workspaceName } = useParams(); // URL 파라미터 추출
